+++ lib/truffle/socket/constants.rb
@@ -73,9 +73,9 @@ class Socket < BasicSocket
   class ResolutionError < SocketError
     attr_reader :error_code
 
-    def initialize(message, error_code)
-      super(message)
+    def initialize(message, error_code = nil)
       @error_code = error_code
+      super(message)
     end
   end
 end